refactor config values management
This commit refactor the config values management to be more clear when
values are filled via command line flags, configuration file or using
environment variables and what are the default values.
The `config.IConfig` interface expose getters and setter to read and set
values on config instance.  These getters are confuse, because some of
them return the value informed by user or the default value and others
just return the value informed by user, which make difficult to inspect
all default values from config. These setters are used to override the
config values from configuration file and environment variables, but they
are confuse because in multiples places we are changing the configuration
that can leave in a inconsistent state. For example, the start and
generate command, call functions to override the config values to the
values from configuration file and environment variables.
So, this commit improve the config management to load the default
values, read the user input from command lines, configuration file and
environment variable in on single place, and change the commands to
receive a "ready to use" config. This commit also remove the
`config.IConfig` interface to make more simple.
NOTE: We still have getters and setters on config struct because we call
theses functions in several places, so we need to remove them and change
the references to use the public fields from config.

Co-authored-by: Matheus Alcantara <matheus.alcantara@zup.com.br>